refactor: reorganize project layout
