+++ .env.example
@@ -1,22 +1,28 @@
-# Movies API E2E Test Environment Configuration
-# Copy this file to .env and customize the values for your environment
+# Repository-wide environment template
+# Copy this file to .env in the repository root if you want shell helpers to
+# pick up default values, then sync the same values into app/.env when running
+# docker compose.
 
-# Service Configuration
+# --- Service configuration ---
 PORT=8080
+AUTH_TOKEN=replace-with-static-token
+DB_URL=postgres://cinema:cinema@db:5432/cinema?sslmode=disable
+BOXOFFICE_URL=https://mock.apifox.com/m1/4288164-0-default
+BOXOFFICE_API_KEY=replace-with-real-api-key
+
+# Optional override for scripts that talk to the running API
 BASE_URL=http://127.0.0.1:8080
 
-# Authentication
-AUTH_TOKEN={{YOUR_STATIC_BEARER_TOKEN_HERE}}
-
-# Database Configuration (for the application, not used directly by e2e tests)
-DB_URL={{YOUR_SELFHOST_DB_URL_HERE}}
-
-# Box Office API Integration
-BOXOFFICE_URL=https://m1.apifoxmock.com/m1/7149601-6873494-default
-BOXOFFICE_API_KEY=0B4nmUwMPBphsKDr_u9HX
+# --- Docker/CI helpers ---
+APP_IMAGE=cinema-app
+POSTGRES_DB=cinema
+POSTGRES_USER=cinema
+POSTGRES_PASSWORD=cinema
+POSTGRES_PORT=5432
+FRONTEND_API_BASE_URL=
 
 # Usage:
 # 1. Copy this file to .env: cp .env.example .env
-# 2. Customize the values in .env for your environment
-# 3. Start your services: make docker-up
-# 4. Run tests: ./e2e-test.sh
+# 2. Copy app/.env.example to app/.env and make sure the values stay in sync
+# 3. Start the stack from the app directory: (cd app && make docker-up)
+# 4. Run the end-to-end tests: (cd app && make test-e2e)
