+++ app/.env
@@ -5,6 +5,7 @@ POSTGRES_PORT=5432
 
 APP_IMAGE=cinema-app
 PORT=8080
+APP_ENV=development
 AUTH_TOKEN=local-token
 
 # 容器内数据库连接串，指向 Compose 服务名 db
@@ -12,6 +13,7 @@ DB_URL=postgres://cinema:cinema@db:5432/cinema?sslmode=disable
 
 # 外部票房 API（可替换为真实地址与密钥）
 BOXOFFICE_URL=https://mock.apifox.com/m1/4288164-0-default
+# OPENAPI_SPEC_PATH=api/specs/service.openapi.yml
 BOXOFFICE_API_KEY=mock-key
 
 # 前端可选的后端访问地址（为空时默认指向当前主机的 8080 端口）
