+++ app/.env.example
@@ -1,0 +1,25 @@
+# --- Application runtime ---
+PORT=8080
+APP_ENV=development
+AUTH_TOKEN=replace-with-static-token
+DB_URL=postgres://cinema:cinema@db:5432/cinema?sslmode=disable
+BOXOFFICE_URL=https://mock.apifox.com/m1/4288164-0-default
+BOXOFFICE_API_KEY=replace-with-real-api-key
+
+# Optional override for serving the bundled OpenAPI spec
+# OPENAPI_SPEC_PATH=api/specs/service.openapi.yml
+
+# --- Docker image configuration ---
+APP_IMAGE=cinema-app
+
+# --- Database bootstrap (used by docker compose) ---
+POSTGRES_DB=cinema
+POSTGRES_USER=cinema
+POSTGRES_PASSWORD=cinema
+POSTGRES_PORT=5432
+
+# External URL exposed through the nginx sidecar (optional)
+FRONTEND_API_BASE_URL=
+
+# Local tooling helper (used by migrations/tests)
+DB_VERIFY_DSN=postgresql://cinema:cinema@127.0.0.1:5432/cinema
