+++ app/internal/boxoffice/client.go
