+++ app/internal/repository/movie_repository.go
@@ -1,7 +1,7 @@
 package repository
 
 import (
-	"cinema/model"
+	"cinema/internal/model"
 	"context"
 	"errors"
 	"time"
