--- cinema/.env.example
@@ -1,11 +1,0 @@
-PORT=8080
-AUTH_TOKEN=local-token
-DB_URL=postgres://cinema:cinema@db:5432/cinema?sslmode=disable
-BOXOFFICE_URL=https://mock.apifox.com/m1/4288164-0-default
-BOXOFFICE_API_KEY=mock-key
-POSTGRES_DB=cinema
-POSTGRES_USER=cinema
-POSTGRES_PASSWORD=cinema
-POSTGRES_PORT=5432
-APP_IMAGE=cinema-app
-DB_VERIFY_DSN=postgresql://cinema:cinema@127.0.0.1:5432/cinema
