+++ scripts/manual_api_smoke.sh
